fix(sensors/gps): unsigned instance index

+++ src/modules/sensors/vehicle_gps_position/gps_blending.hpp
@@ -66,7 +66,7 @@ public:
 	// define max number of GPS receivers supported for blending
 	static constexpr int GPS_MAX_RECEIVERS_BLEND = 2;
 
-	void setGpsData(const sensor_gps_s &gps_data, int instance)
+	void setGpsData(const sensor_gps_s &gps_data, uint8_t instance)
 	{
 		if (instance < GPS_MAX_RECEIVERS_BLEND) {
 			_gps_state[instance] = gps_data;
@@ -78,7 +78,7 @@ public:
 	void setBlendingUseVPosAccuracy(bool enabled) { _blend_use_vpos_acc = enabled; }
 	void setBlendingTimeConstant(float tau) { _blending_time_constant = tau; }
 	void setPrimaryInstance(int primary) { _primary_instance = primary; }
-	void setAntennaOffset(const Vector3f &offset, int instance)
+	void setAntennaOffset(const Vector3f &offset, uint8_t instance)
 	{
 		if (instance < GPS_MAX_RECEIVERS_BLEND) { _antenna_offset[instance] = offset; }
 	}
